requirements.txt
protobuf==6.30.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.35.0
